+++ lib/stemmer.dart
@@ -1,0 +1,3 @@
+library stemmer;
+
+export 'PorterStemmer.dart';