+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - 22/7/2019
+
+- Added unit tests.
+- Improved `README.md`.
+- Added example usage.
+
 ## [1.0.0] - 21/7/2019
 
-- Initial port of the NLTK PorterStemmer implementation
+- Initial port of the NLTK PorterStemmer implementation.